requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
Django==2.1.2
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
